DeleteMaskに関する部分を分離
