+++ Assets/MeshDeleterWithTexture/Editor/Scripts/DeleteMaskCanvas.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 60b12da0ed4af7749b47cfb20f97b691
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
